Consolidate UI control methods in UIControls.cs.
