+++ Changelog.txt
@@ -1,12 +1,14 @@
 - Major overhaul: new volumetric population calculations
 - Major new feature: selectable configuration presets for each service type
 - Major new feature: optional realistic student capacity calculations for elementary and high schools
-- New feature: adjust crime rate
+- Major new feature: adjust crime rate
 - Rearrange options panel layout and add sub-tabs as required
+- Enable preview of calculated building floors in building preview
 - Add custom mouse button keybinding option
 - Upgrade to Harmony 2
 - Add settings filters to building details panel
 - Backport building preview renderer improvements from Ploppable RICO Revisited
+- Multiple code cleanups
 
 
 Version 1.4.3 -
