+++ Code/Settings/CalculationTabs/PopulationPanel.cs
@@ -54,11 +54,11 @@ namespace RealisticPopulationRevisited
             rowLabels = new UILabel[5];
 
             // Service selection dropdown.
-            serviceDropDown = PanelUtils.AddPlainDropDown(panel, Translations.Translate("RPR_OPT_SVC"), serviceNames, -1);
+            serviceDropDown = UIControls.AddPlainDropDown(panel, Translations.Translate("RPR_OPT_SVC"), serviceNames, -1);
             serviceDropDown.parent.relativePosition = new Vector3(20f, Margin);
 
             // Pack selection dropdown.
-            packDropDown = PanelUtils.AddPlainDropDown(panel, Translations.Translate("RPR_OPT_CPK"), new string[0], -1);
+            packDropDown = UIControls.AddPlainDropDown(panel, Translations.Translate("RPR_OPT_CPK"), new string[0], -1);
             packDropDown.parent.relativePosition = new Vector3(20f, PackMenuY);
 
             // Headings.
@@ -93,7 +93,7 @@ namespace RealisticPopulationRevisited
 
             // Pack name textfield.
             packNameField = UIUtils.CreateTextField(panel, 200f, 30f);
-            UILabel packNameLabel = PanelUtils.AddLabel(packNameField, Translations.Translate("RPR_OPT_EDT_NAM"), -100f, (packNameField.height - 18f) / 2);
+            UILabel packNameLabel = UIControls.AddLabel(packNameField, Translations.Translate("RPR_OPT_EDT_NAM"), -100f, (packNameField.height - 18f) / 2);
             packNameField.relativePosition = new Vector3(140f, currentY);
             packNameField.padding = new RectOffset(6, 6, 6, 6);
             packNameField.isEnabled = false;
